refactor(test): move sourcemap util functions to testUtil
